refactor: Add core dir (#1)
