+++ src/user/models.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, String, BigInteger, DateTime
 from sqlalchemy.sql import func
-from database import Base
+from core.database import Base
 
 class User(Base):
     __tablename__ = "user"
